requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
aiofiles==24.1.0
openai==1.3.7
spacy==3.7.2
httpx==0.25.2
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4